--- hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/hudi/functional/SecondaryIndexTestBase.scala
@@ -1,130 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.hudi.functional
-
-import org.apache.hudi.DataSourceWriteOptions.{HIVE_STYLE_PARTITIONING, PARTITIONPATH_FIELD, PRECOMBINE_FIELD, RECORDKEY_FIELD}
-import org.apache.hudi.common.config.{HoodieMetadataConfig, TypedProperties}
-import org.apache.hudi.common.model.FileSlice
-import org.apache.hudi.common.table.HoodieTableMetaClient
-import org.apache.hudi.config.HoodieWriteConfig
-import org.apache.hudi.metadata.HoodieMetadataFileSystemView
-import org.apache.hudi.testutils.HoodieSparkClientTestBase
-import org.apache.hudi.util.JFunction
-import org.apache.hudi.{DataSourceReadOptions, HoodieFileIndex}
-
-import org.apache.spark.sql.catalyst.expressions.{AttributeReference, EqualTo, Expression, Literal}
-import org.apache.spark.sql.types.StringType
-import org.apache.spark.sql.{DataFrame, SparkSession}
-import org.junit.jupiter.api.Assertions.assertTrue
-import org.junit.jupiter.api.{AfterEach, BeforeEach}
-
-import scala.collection.JavaConverters
-
-class SecondaryIndexTestBase extends HoodieSparkClientTestBase {
-
-  var spark: SparkSession = _
-  val metadataOpts: Map[String, String] = Map(
-    HoodieMetadataConfig.ENABLE.key -> "true",
-    HoodieMetadataConfig.RECORD_INDEX_ENABLE_PROP.key -> "true"
-  )
-  val commonOpts: Map[String, String] = Map(
-    "hoodie.insert.shuffle.parallelism" -> "4",
-    "hoodie.upsert.shuffle.parallelism" -> "4",
-    HoodieWriteConfig.TBL_NAME.key -> "hoodie_test",
-    RECORDKEY_FIELD.key -> "record_key_col",
-    PARTITIONPATH_FIELD.key -> "partition_key_col",
-    HIVE_STYLE_PARTITIONING.key -> "true",
-    PRECOMBINE_FIELD.key -> "ts"
-  ) ++ metadataOpts
-  var mergedDfList: List[DataFrame] = List.empty
-
-  @BeforeEach
-  override def setUp(): Unit = {
-    initPath()
-    initQueryIndexConf()
-    initSparkContexts()
-    initHoodieStorage()
-    setTableName("hoodie_test")
-    spark = sqlContext.sparkSession
-  }
-
-  @AfterEach
-  override def tearDown(): Unit = {
-    cleanupFileSystem()
-    cleanupSparkContexts()
-  }
-
-  def verifyQueryPredicate(hudiOpts: Map[String, String], columnName: String): Unit = {
-    mergedDfList = spark.read.format("hudi").options(hudiOpts).load(basePath).repartition(1).cache() :: mergedDfList
-    val secondaryKey = mergedDfList.last.limit(1).collect().map(row => row.getAs(columnName).toString)
-    val dataFilter = EqualTo(attribute(columnName), Literal(secondaryKey(0)))
-    verifyFilePruning(hudiOpts, dataFilter)
-  }
-
-  private def attribute(partition: String): AttributeReference = {
-    AttributeReference(partition, StringType, nullable = true)()
-  }
-
-
-  private def verifyFilePruning(opts: Map[String, String], dataFilter: Expression): Unit = {
-    // with data skipping
-    val commonOpts = opts + ("path" -> basePath)
-    metaClient = HoodieTableMetaClient.reload(metaClient)
-    var fileIndex = HoodieFileIndex(spark, metaClient, None, commonOpts, includeLogFiles = true)
-    val filteredPartitionDirectories = fileIndex.listFiles(Seq(), Seq(dataFilter))
-    val filteredFilesCount = filteredPartitionDirectories.flatMap(s => s.files).size
-    val latestDataFilesCount = getLatestDataFilesCount(opts)
-    assertTrue(filteredFilesCount > 0 && filteredFilesCount < latestDataFilesCount)
-
-    // with no data skipping
-    fileIndex = HoodieFileIndex(spark, metaClient, None, commonOpts + (DataSourceReadOptions.ENABLE_DATA_SKIPPING.key -> "false"), includeLogFiles = true)
-    val filesCountWithNoSkipping = fileIndex.listFiles(Seq(), Seq(dataFilter)).flatMap(s => s.files).size
-    assertTrue(filesCountWithNoSkipping == latestDataFilesCount)
-  }
-
-  private def getLatestDataFilesCount(opts: Map[String, String], includeLogFiles: Boolean = true) = {
-    var totalLatestDataFiles = 0L
-    val fsView: HoodieMetadataFileSystemView = getTableFileSystemView(opts)
-    try {
-      fsView.getAllLatestFileSlicesBeforeOrOn(metaClient.getActiveTimeline.lastInstant().get().getTimestamp)
-        .values()
-        .forEach(JFunction.toJavaConsumer[java.util.stream.Stream[FileSlice]]
-          (slices => slices.forEach(JFunction.toJavaConsumer[FileSlice](
-            slice => totalLatestDataFiles += (if (includeLogFiles) slice.getLogFiles.count() else 0)
-              + (if (slice.getBaseFile.isPresent) 1 else 0)))))
-    } finally {
-      fsView.close()
-    }
-    totalLatestDataFiles
-  }
-
-  private def getTableFileSystemView(opts: Map[String, String]): HoodieMetadataFileSystemView = {
-    new HoodieMetadataFileSystemView(metaClient, metaClient.getActiveTimeline, metadataWriter(getWriteConfig(opts)).getTableMetadata)
-  }
-
-  private def getWriteConfig(hudiOpts: Map[String, String]): HoodieWriteConfig = {
-    val props = TypedProperties.fromMap(JavaConverters.mapAsJavaMapConverter(hudiOpts).asJava)
-    HoodieWriteConfig.newBuilder()
-      .withProps(props)
-      .withPath(basePath)
-      .build()
-  }
-
-}
